8_News_Research_Assistant/requirements.txt
openai>=1.12.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
newsapi-python>=0.2.6
python-dateutil>=2.8.2
rich>=10.0.0